changed userInfo type to NSDictionary*
git-svn-id: svn://orca.physics.unc.edu/Orca@11838 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/IPE/IPE SlowControl/ORIpeSlowControlController.h
@@ -141,7 +141,7 @@
 #pragma mark •••Data Source Methods (OutlineView)
 - (int) outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item;
 - (BOOL) outlineView:(NSOutlineView *)outlineView isItemExpandable:(id)item; 
-- (id) outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item; 
+- (id) outlineView:(NSOutlineView *)outlineView child:(NSUInteger)index ofItem:(id)item; 
 - (id) outlineView:(NSOutlineView *)outlineView objectValueForTableColumn:(NSTableColumn *)tableColumn byItem:(id)item;
 
 #pragma mark •••Data Source Methods (ComboBox)
